simplify stream play on card component
